Add binary test application & move ga examples
