+++ example/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "example"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+ecrs = {path = "../" }
+itertools = "0.10.5"
+rand = "0.8.5"
