+++ example/src/main.rs
@@ -1,0 +1,6 @@
+#![allow(dead_code)]
+mod ga;
+
+fn main() {
+	ga::ga_example();
+}
